services/app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-logging==0.48b0
opentelemetry-semantic-conventions==0.48b0
python-multipart==0.0.20
python-dotenv==1.0.1
requests==2.32.2
opentelemetry-api
